api/langchain_implementation/requirements.txt
langchain==0.0.267
# streamlit==1.25.0
transformers==4.31.0
torch==2.0.1
einops==0.6.1
bitsandbytes==0.41.1
accelerate==0.21.0
pdfminer.six==20221105
bs4==0.0.1
sentence_transformers
# duckdb==0.7.1
chromadb==0.3.26
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
requests==2.31.0
uvicorn==0.18.3
# torch==2.0.1
torchvision==0.15.2
openai
huggingface_hub
python-dotenv
langserve[all]